requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
flask-login==0.6.3
flask-migrate==4.0.5
flask-admin==1.6.1
Flask-Mail==0.9.1
Flask-Session==0.4.0
flask-cors==6.0.2
flasgger==0.9.7b2
pymongo==4.6.1
celery==5.3.6
redis==5.0.1
requests>=2.31.0
beautifulsoup4==4.12.2
Pillow
pydantic[email]
authlib
python-dotenv==1.0.1
langchain_classic==1.0.1
langchain-community
langchain_core
langchain_google_genai
faiss-cpu
gunicorn
numpy==1.26.4
packaging
PyJWT